More tweaks
This change does two more things that we discussed. It's a bit messy and you might want to tweak it further, but the change I made was to swap ol for ul so that we can explain the 1., 1., 1., lists after we "gussy" things up. 

I also added a mention to the fact that this is written in MD towards the end.